+++ Flepper.QueryBuilder/Utils/Extensions/ColumnsExtensions.cs
@@ -6,6 +6,7 @@ namespace Flepper.QueryBuilder.Utils.Extensions
     {
         private const string ALIAS = " AS ";
         private static readonly string[] AliasSplitter = { ALIAS, " as " };
+
         public static string JoinColumns(this IEnumerable<string> columns)
             => string.Join(",", columns) + " ";
     }
